fix bugs with Higgs masses in DataMCType and variable types in HistoConfig and Makefile

+++ Code/HistoConfig.cxx
@@ -212,7 +212,7 @@ bool HistoConfig::hasID(int64_t id_){
 }
 
 
-int HistoConfig::GetID(unsigned int i){
+int64_t HistoConfig::GetID(unsigned int i){
   if(ID.size()>i) return ID.at(i);
   return -999;
 }
